Turn System
Turn system implemented, Everything is setup so we can polish if needed.
Basically it works like this:
The game sets up itself(we run animation/telegraph if we have).
Then it's player turns and later enemy turn.
If player has won or lost, the state will be changed to won or lost and we will then enable other things(still yet to be done)